is_sheik: modified get_frame_data as list comp for a minor preformance gain. Collect all frames after the first 5 seconds. Save the file.
feature_pruning:
deleted ['DPAD_LEFT', 'DPAD_RIGHT', 'DPAD_DOWN', 'DPAD_UP', 'START'] from
the numpy array in 'TimeSeries'
Split the columns of timeSeries into each input.
Added a column of the max of X and Y
Added a column that is the max of the triggers
added polar coordinates of the J-stick and C-stick.