Added a file to train a model with the parameters copied and pasted from an optuna trial. I have one for mixed precision and one without.
The mixed precision is 33% faster to train,
but it did not seem to converge as fast.
I should test to see if training for the same
time results in similar results. I have not
checked the memory consumption of mixed precision.

The results of training the model with the best
optuna parameters was really bad. We really overfit.
I don't know why that would be. I did all this with
ChatGPT and have not checked over it. Maybe there is
some mistake.

It might be that we have to be more careful about
optuna getting really lucky with the validation set.